BID API: create-user now handles DB integrity errors better
Instead of just returning a 500 Internal Server Error, we check the type
of integrity error, and if it's something to do with the email address,
we act as if it's a duplicate error. This is a fairly safe assumption,
since that's the only thing that we can get a race condition on -- the
other things (length, for example) are checked by the form already.