Make it possible to turn off the flatpages app
This app catches all 404 errors and tries to look up a flatpage that
matches the URL. If this can also not be found, it returns a 404 itself.
When DEBUG=True this behaviour blocks Django's default behaviour of showing
the list of URL rules, so while debugging it is nice to have the flatpages
app disabled.