Tweaked CSRF handling on login page
The view now first checks whether you've already logged in. If this is the
case, it'll just redirect you to the success URL without checking the
CSRF token. This should fix the case where someone is looking at an old
login form while already logged in on Blender ID in another tab.

If the user is not logged in yet, the superclass implementation is called,
which checks the CSRF token anyway.