Gracefully handle DB integrity error when confirming email
This can happen when changing the email address to one that's free at the
time of filling out the form, but already in use when following the
confirmation link.

There could be other reasons why this happens, which is why I added a bit
more elaborate logging to allow better debugging.