Added nickname field to user model, including migrations and API
Be sure to run all migrations in one go, as multiple migrations were
needed to get this to work:

- 0016: Adds an optional 'nickname' field to the User model.
- 0017: Iterates over all users with NULL nickname, and assigns a unique
        one.
- 0018: Modifies the User model to make 'nickname' non-optional and
        unique, so that the database guards the uniqueness constraint.

Note that rolling back migration 0017 erases the username for all users.
If any usernames were set between the execution of 0016 and 0017, those
will be gone after rolling back 0017.

The API call for user creation doesn't use a nickname field; it creates
its own nickname. This makes it easier to deploy, as the API itself
doesn't change (otherwise we'd have to add nicknames to the Blender
Store).