Migrated from requirements.txt to pipenv
On the production server we use PIPENV_VENV_IN_PROJECT=1 so that the
virtualenv is created in a predictable location.

By moving to pipenv we automatically got some upgraded dependencies. The
only one that caused an issue was django-oauth-toolkit (1.0.0 → 1.1.2)
As they changed their abstract model we needed to create a migration to
handle that for our own concrete subclasses.

Django is still held back at <2.0, which means django-oauth-toolkit is
held back at <1.2.0.