Extended Badge API with a HTML endpoint at /api/badges/<user-id>/html
Optionally the URL can be extended with `/<size>` to indicate a size in
{'s', 'm', 'l'}. This set can be extended/refined later.

Passing the user ID isn't strictly necessary. However, it does allow
for caching the result (for a short while) based on the URL alone,
and without having to cache the authentication token too. This seemed like
a good idea to me.

This adds a dependency: sorl.thumbnail. It handles automatic on-demand
resizing of uploaded images in an `ImageField`. Note that the README.md has
been updated for this, most notably with an extension of the cron job to
clean up obsolete cached images.