Added django-admin-select2 to make editing of user roles more Ton-proof
With the default <select> box, giving someone a role by clicking on that
role can revoke roles that are not visible (the entire list of roles
doesn't fit in the box and requires scrolling). This makes it easy to make
accidental mistakes.

By switching to select2, granting and revoking roles can be done by clicking,
without affecting non-clicked roles.