Logout: allow redirecting to another site via /logout?next=some-url
The new URL must be to the same host as the HTTP `Referer` header. Without
that header, or with a different host, the user will be redirected to
the `/about` page. Furthermore, downgrading from HTTPS to HTTP is not
allowed (if the referer was using HTTPS, you cannot redirect to HTTP).