Self-hosted avatars for users
Users can upload images (up to 2 MB), which will be converted to JPEG,
cropped square, and resized to 160x160 pixels upon first use. Since the
first use is for showing their profile after they edited it, this is
pretty much immediately.

A request to /api/users/{user ID}/avatar either redirects to an image/jpeg of
the 160x160 pixel avatar, or it redirects to the default
image/png avatar. The URL is publicly reachable.

It is up to the caller to decide what to do with the redirect to the
default avatar (download & use the default image as the user's avatar,
or mark the user as "without avatar").